+++ tests/test_classes.cpp
@@ -575,7 +575,7 @@ NB_MODULE(test_classes_ext, m) {
     struct Int {
         int i;
         Int operator+(Int o) const { return {i + o.i}; }
-        Int operator-(float j) const { return {int(i - j)}; }
+        Int operator-(float j) const { return {int((float)i - j)}; }
         bool operator==(Int o) const { return i == o.i; }
         Int &operator+=(Int o) {
             i += o.i;
@@ -916,13 +916,13 @@ NB_MODULE(test_classes_ext, m) {
         .def_ro("value", &NewDflt::value);
     nb::class_<NewStarPosOnly>(m, "NewStarPosOnly")
         .def(nb::new_([](nb::args a, int value) {
-            return NewStarPosOnly{nb::len(a) + value};
+            return NewStarPosOnly{nb::len(a) + (size_t) value};
         }),
             "args"_a, "value"_a = 42)
         .def_ro("value", &NewStarPosOnly::value);
     nb::class_<NewStar>(m, "NewStar")
         .def(nb::new_([](nb::args a, int value, nb::kwargs k) {
-            return NewStar{nb::len(a) + value + 10 * nb::len(k)};
+            return NewStar{nb::len(a) + (size_t) value + 10 * nb::len(k)};
         }),
             "args"_a, "value"_a = 42, "kwargs"_a)
         .def_ro("value", &NewStar::value);
